requirements.txt
# Core ML / NLP
sentence-transformers>=2.7
faiss-cpu>=1.8
numpy>=1.26
scikit-learn>=1.4

# LLM
openai>=1.30
tiktoken>=0.7

# Web scraping
requests>=2.32
beautifulsoup4>=4.12
lxml>=5.2

# Data
pandas>=2.2

# API
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
httpx>=0.27

# Config
python-dotenv>=1.0
